Fix meta refresh test
This fixes the less-than ideal test for <meta> refresh I wrote.
I don't think it needs a long timeout, but just had to wait for
the iframe to load completely, which is when, as per the spec, the
refresh timer should start.
(Single, tiny HTML file in this case, really shouldn't take long but I
guess the browser is pretty busy when all the tests are run at once?)
I can only go off of my local test results for this currently
but I hope that this fixes the failures that came up.

Spec:
Step 13 of the shared declarative refresh steps:
https://html.spec.whatwg.org/#shared-declarative-refresh-steps

Bug: 333976551
Change-Id: I134348aac4f97b321875b37a7c179432585bd3fc