Add support for calc-size() and interpolate-size to flex-basis.
This contains both the changes to support calc-size() and
interpolate-size on flex-basis, and the changes to support the content
keyword as the basis of a calc-size() expression.

There doesn't appear to be any code specifically testing for the
presence of the 'content' value, so we don't need to add a
blink::Length::HasContent() method and the code that would support it.

Bug: 313072, 353538495
Change-Id: Ic2f97688fbe78b47a53ee6c1559ddc4dea185729